// src/app/socket/events/events.enum.ts
export enum Events {
    Connection = 'connection',
    Disconnect = 'disconnect',
    Welcome = 'welcome',
    DataRegistration = 'data-registration',
    SuccessfulPairing = "successful-pairing",
    PartnerDisconnected = "partner-disconnected",
    SearchStart = "search-start",
    SearchStop = "search-stop",
    SendMessage = "send-message",
    MessageReceived = "message-received",
    LeaveCall = "leave-call"
}